Require explicit args to DictBasedArgument
This restricts the set of allowed args that can be passed
to DictBasedArgument.  This makes it easier to read because
rather than taking a dictionary we specify *exactly* what
arguments we accept.

Also removed metavar from the model.  This is only used for
display purposes which bcdoc does not look at.