Fix memory utilization for uploads
A ReadFileChunk is introduced so that we can upload file parts
from a file rather than buffering the entire part in memory.

Also, the etag verification for downloads was chnaged. What we do
instead is calculate the md5 writing the file, and if the md5
verification fails, we delete the file.