Remove src/dest resource existence check
Rather, just try the operation and handle the case
when it fails.  This is to address issue 328.  The problem
is that in order to verify that the resource exists, we need
a separate set of API calls than we use to actually retrieve/upload
the resource.  This is problematic if you want to give specific
permissions for a user (they only have access to download files, they
can't see a list of available files).