Honor REQUESTS_CA_BUNDLE env var
This allows a user to specify a different cert bundle via an
env var.  This uses the new verify arg added to get_endpoint
in botocore.

Previously it was creating an Endpoint object and then setting the
verify attribute, but the REQUESTS_CA_BUNDLE lookup happens before
an Endpoint object is created, so we would override the value
from the env var.  By using the verify arg in get_endpoint we
can avoid this.