Move test utilities into awscli.testutils module
Given that we do not include the tests/ directory as part of the awscli
package, external code that wants to use the test utilities we provide
has to copy/paste the code into their own package.

In order to avoid this, all the test utilties have been moved over to a
testutils module in the CLI.  Only test code is expected to import this
module.  I've updated all our unit/integ tests to import from here.