Register timestamp parser after profile is set
The issue was that setting a profile for a session resets all
the registered components for the session, including the
response parser that alters the timetsamp parsing behavior in the
AWS CLI.

To fix this, I've added a new event that I think will be useful
in general, a `session-initialized` event, sent once the session
has its profile set.  There are actually several handlers that
I think should switch over to this signal, which I'll add in
a separate pull request.

I've also added an integration test to ensure we don't regress
on this issue again.