Parse --statistic-values values into a decimal type
This was failing the improved botocore param validation because
we were previously setting this value to a string.  We now set
this value to a decimal.Decimal to ensure we have a valid type
that will also pass the botocore param validation.

While I was in the code, I also updated the code to use the new
botocore CamelCasing from metric_data to MetricData.  There was
also some general test cleanup to switch the tests to
assert_params_for_cmd2.

Fixes #1036.