Move part size constants out to their respective modules
As part of refactoring out the runtime config, the constants
that _should_ not change (because they're based on S3 limits)
are moved to the corresponding modules that use these values.

This makes a clear distinction between values tied to s3 limits,
and values the user can change if needed.

Also, based on the latest S3 docs, the maximum number of parts
is now 10000 instead of 1000 so I've updated that accordingly.
This resulted in a few unit tests updates.